GUI: mux: only remove auto-added uniqueness suffixes from destination file name
Removing all suffixes of the form "a space character followed by (
followed by a number followed by )" would also remove things such as
years (e.g. "Amélie (2001)" was truncated to "Amélie" before trying to
make the file name unique).

Now the GUI remembers which suffix it hat to add the last time it
tried to make the name unique, and only that suffix will be removed on
the next attempt.

Fixes #2521.