AUTHORS: don't enforce <last name, first name> format
A lot of cultures do not adhere to this type of format. There doesn't
have to be a last name in the first place, or a first one. The family
name doesn't have to be the last one, it might be the first one. There
might be a lot more components to names.