GUI: jobs: add action to execute type "delete source files"
This is a destructive action; no user confirmation is requested. The
action only works for multiplexer jobs, not for info jobs.

The action is disabled by default. It has to be enabled to be usable
at all, and if it should always be executed, one of the two options
"After a job completes successfully or with warnings" or "After a job
exits with an error or the user aborts the job" must be enabled, too.

Implements #869 and #2570.