GUI: default for often used languages from Wikipedia's languages by native speakers
The prior selection was pretty Western-centric with a couple of
additions from Asia. However, there are so many more languages, and
for a lot of them there are many more native speakers than for those
which were on the list before this change.

The new list simply mirrors Wikipedia's "List of languages by number
of native speakers" found at
https://en.wikipedia.org/wiki/List_of_languages_by_number_of_native_speakers

Additionally the GUI will only show the often used languages in
drop-down boxes by default.

Both can still be changed in the preferences.