GUI: update often used languages to v37 built-ins if not modified by user
The intention in v37 was to update both the setting "only include
often used entries" and the list of often used languages at the same
time, as long as the user hadn't modified that list.

Unfortunately only the checkbox's default value was updated, but not
the list. This lead to the confusing situation that suddenly only ~15
languages were shown in the drop-down boxes for users upgrading their
installation, such as the person in #2611.

New installations were not affected.