AAC reader: use probe range info from file type detection stage
There are files that start with garbage, and where that garbage
contains data that forms valid ADTS or LATM headers. In that case the
file type detection might detect the valid data (more than e.g. eight
consecutive valid headers of the same type with the same parameters)
starting at an offset, e.g. at position 100.000.

However, the code reading the file headers later on only looked for
the first header that passed validation, which might mean a header at
a much earlier position in the file — which only looks like a valid
header by accident. The result was wrong codec parameters.

Fixes #2622.