mkvmerge: refactor file probing/header reading code
Before the file probing code opened the file once, did its thing, and
closed the file again. No information from that stage was available to
the header reading stage. This meant that headers had to be parsed
twice for several formats. The probing function was a static member
function, making passing information to the header reading function,
which was an instance member function, even more difficult.

After refactoring the file is only opened once. The same open file
handle is used by both the probing & the header reading
functions. Both of those functions are now instance member functions,
and the same object instance is used for both calls, allowing passing
information from the probing to the header reading stage.

Additionally this enabled the removal of a lot of boilerplate code
such as passing variables to the various probing functions or handling
exceptions in each probing function instead of in the function calling
the probes. The git diff stats confirm this rather impressively:

78 files changed, 769 insertions(+), 1778 deletions(-)

A net gain of more than 1000 lines.