MPEG TS: only consider packets actually copied for min timestamp calculation
MPLS playlists contain restrictions which content to copy from an M2TS
file. mkvmerge does follow those restrictions when copying the
packets, but not when determining the minimum timestamp in the which
will be used for shifting the output timestamps down to 0. This meant
that content not actually copied might determine the minimum
timestamp.

As a result the minimum timestamp subtracted from all other timestamps
might be too small (if there's content before the start of the MPLS's
timestamp restriction).

Fixes #2670.