GUI: mux: ensure "playlist scan finished" signal is handled
If the connection for the signal is made while the identifier is
running, the corresponding signal might already have been fired
leading to it not being handled at all.

By connecting the signal before the worker thread is started, no
signals will get lost when input files are playlists & scan very
quickly.

Fixes #2678.