GUI: mux: let user select which disc library info to use
This extends the "select playlist to add" dialog with a list of all
disc library entries found. The user can select which one to add
instead of the GUI's default. The GUI still defaults to the English
one if found or the first one otherwise.

This is the real fix for #2765.