GUI: mux: let Qt track signal recipient
Using the three-argument variant of the `connect()` call means that Qt
doesn't have an object to track; it just calls a functor. If that
functor uses a pointer to an object, then removing said object means
that the functor will still be called and access a pointer to invalid
memory.

Using the four-argument variant instead allows Qt to automatically
remove the connection when the object in question is removed.

This happens in the case that a multiplexer tab is opened (e.g. on
program startup) closed, and later the preferences are changed.

Fixes #2785.