GUI: change default list of "often used languages"
The previous default list was derived from the list of spoken
languages sorted by native speakers. This list was so long that in
addition the option "only show often used languages" was enabled so as
not to confuse users with very long drop-down boxes where the top list
was very long.

Unfortunately this resulted in users being confused because their
particular language wasn't present in the drop-down boxes, and they
didn't know that this is a feature that can be customized.

The default settings matter a lot for user experience; most users
rather assume the program's buggy than they can configure a specific
feature to work as they expect.

The new default is to only include the following entries at the
top and all the other available languages below (meaning "only show
often used languages" is off by default again):

- mul: multiple languages
- zxx: no linguistic content
- qaa: reserved for local use
- mis: uncoded languages
- eng: English

Additionally the system's default user interface language will be
included, too.

Implements #2716.