MPEG TS reader: MPLS: accept all packets once play item's minimum timestamp was encountered
An MPLS playlist consists of one or more play items. Each play item
has a start and end timestamp. The intended use of the start
timestamp: locate the first packet whose timestamp is equal to or
larger than the play item's start timestamp. Demux that and all
following packets.

Unfortunately mkvmerge interpreted the timestamp differently. It would
compare the timestamps of all packets to the play item's start
timestamp. This is problematic when the timestamps aren't strictly
monotonic, which they often aren't.

For example, on a certain Blu-ray the timestamps for a PGS subtitle
track would look like this:

1. PTS 00:02:46.054911111 PGS type 22 (Presentation Composition)
2. PTS 00:02:46.050422222 PGS type 23 (Window Definition)
3. PTS 00:02:46.041455555 PGS type 20 (Palette Definition)
4. PTS 00:02:46.050422222 PGS type 21 (Object Definition)
5. PTS 00:02:46.050422222 PGS type 128 (End Of Display)

With the play item's start timestamp at 00:02:46.054911111 mkvmerge
would drop packets 2–5 and only keep 1 — destroying the PGS entry in
the process.

Additionally this resulted in the timestamps of the following PGS
subtitle being wrong: it would use the timestamp of the garbled first
entry instead.

Part of the fix of #2824.