Matroska reader: fix `default track` flag handling
Let's take a Matroska file with a subtitle track with its `default
track` set to `on`. When you remux that file, forcing that flag to off
via `--default-track …:no`, the first other track of the same
type (subtitles) that is read from a container which does not provide
such `default track` information (e.g. an SRT file) should have its
`default track` flag set.

Before this change this wasn't the case; the track from the SRT would
have its flag set to `no`. The Matroska reader was always signalling
"default track = yes` at the container level to the packetizer, even
if it was overridden via command line options. The flag would still be
set to `no` in the output file, but the fact that there was a `yes` at
the container level prevented `mkvmerge` from considering other tracks
which only had a priority of `track type`.

Fixes #2863.