mkvmerge man page: remove section about subtitles
First of all, the section of supported formats was incomplete/out of
date. Second, I don't include similar lists for supported audio and
video formats either. Last, the question of whether or not a
codec ("format") is supported also depends on the container format it
is in, making such a list questionable.

Fixes #2878.