MP4 reader: drastically reduce number of memory allocations for DASH files
In DASH files there are a lot of fragments (`moof` atom and its
children). Especially the `trun` atom handling was prone to require a
lot of memory re-allocation. Now memory is only re-allocated in larger
chunks, cutting down the header parsing time for one particular 500 MB
DASH file from 5s down to 0.5s.

Part of the fix of #2900.