BCP 47: format includes all components even if language is empty
Before there was a shortcut: if language is empty, only private use
was formatted as that's the only possible way to produce a valid
tag. However, the GUI uses formatting of potentially invalid tags and
then parsing them again as a test for validity. Therefore the
formatting function must always format all individual components.

Fixes #2924.