common: initialize exe directory first, fixing gettext 0.21 on mingw
gettext 0.21 caches opened catalogs differently than prior versions
did, at least on mingw, where it handles directories as wide-char
strings instead of char-strings.

MKVToolNix always initializes the English locale first. The
command-line parsers are the ones that call `gettext()` before they
start parsing the command-line arguments. That happens right in the
common initialization/setup routine.

For `gettext()` initialization MKVToolNix must provide the path to the
catalogs. For Windows it concatenates the installation path (meaning
the directory the current executable is located in) with the
sub-directory `locales`.

During that very first initialization the path to the current
executable hasn't been determined yet, though, resulting in the
relative path `locales` to be passed to
`bindtextdomain()`. `gettext()` then locates the English catalog,
finds it, opens it & caches its location.

After parsing the command-line arguments MKVToolNix might need to
change the locale (e.g. due to the `--ui-language` argument or, in
case of the GUI, due to values stored in its configuration). At that
point the exe location has been determined. Therefore the next call to
`bindtextdomain()` will receive an absolute path, not a relative one.

For some reason I wasn't able to determine this leads to the lookup of
cached catalogs coming to the (wrong) conclusion that the catalog's
been opened already, even though the requested locale is
different (e.g. `de` vs. the initial `en`). No matter, the
already-opened catalog for English is used.

This patch changes determining the path to the exe to be the first
action to be taken in the common initialization/setup routine. That
way the first invocation of `bindtextdomain()` will also be done with
an absolute path; in fact, all calls to `bindtextdomain()` will now be
done with the same absolute path.

The result is that the cache lookup inside `gettext()` comes to the
right conclusion that the catalog for locales other than English has
in fact not been opened yet. It does another lookup, this time for the
new locale, and everything works as expected.