mkvmerge: HEVC: normalize parameter set NALUs
Normalization means that each key frame is prefixed with exactly one
set of the currently active picture, sequence & video parameter sets.

Fixes #3034 and prevents the same class of issues where the parameter
sets required to decode do not match the ones in CodecPrivate but
aren't present in front of the key frames. This often happens when
appending files.