generic_video_packetizer_c: handle broken B frames
Some programs (e.g. makemkv) seem to produce files with only BlockGroups
but no SimpleBlocks at all. However the BlockGroups consists of B frames
that have only forward ReferenceBlock but not a backward one. It appears
that we need to manually set their backward reference to the latest
P frame found for mkvmerge to have the video stream remuxed properly.

I have not tested with a lot of samples, but I did test thoroughly to
confirm it is the right thing to do, namely re-remux the remux output
to make sure they have consistent timestamps and references, cross-
check those with an output muxed with its raw ES stream (which
was carefully extracted / cross-checked).

Extraction without --raw will give proper / identical streams now.

Note that m_ref_timestamp will need to track both I and P frames as
it has been doing for this to work. Not sure why but it is the fact.