HEVC parser: clone memory when retaining parameter sets
When reading HEVC data from certain container types such as Matroska
the memory the parser parses does not belong to the parser. Therefore
all memory the parser needs to retain beyond the end of the current
call to `parse()` must be cloned.

This affects the parameter sets (SPS, PPS & VPS) which are retained
for prefixing key frames with them.

Fixes #3083.