HEVC packetizer: calculate precise frame duration
When the timestamps & the default duration comes from a container with
low timestamp resolution (= from Matroska), calculating the actual
default duration and frame duration to use requires using the HEVC
stream's frame duration if the delta between it and the source's
imprecise duration is less than the source's timestamp resolution.

Otherwise the cluster helper will often come to the conclusion that a
frame's duration doesn't match the track's default duration. The
result in such cases is that `BlockGroup` elements with
`BlockDuration` elements have to be used to signal the frame's (wrong
because imprecise) duration.

With precisely calculated frame duration the cluster helper will
always see that the frame's duration matches the track's default
duration. It can therefore always use `SimpleBlock` elements and save
on space that way.

Fixes #3114.