GUI: multiplexer: overhaul "default track" flag handling
The Matroska specs have recently clarified regarding the intended
interpretation of the "default track" flag. If it is set, it is
supposed to signal to the player that this track is eligible for being
played by default, potentially taking other factors such as user
preferences regarding languages into account.

This implies that more than one track of each type can have this flag
set. For example, a Blu-ray disc with three audio tracks might have
the main audio in both English and Japanese, whereas the third audio
track contains the director's comments. In such a case the first two
tracks should have the "default track" flag set, the third one
shouldn't.

Earlier mkvmerge was enforcing that only one track of each type could
have the flag set. This restriction has been removed.

The GUI has been updated to allow setting the flag for multiple
tracks, too.

Part of the implementation of #3092.