extract: check both bref and fref to determine whether a frame is keyframe
Some programs (e.g. makemkv) seem to produce files with BlockGroup that
consist of B frame and only forward ReferenceBlock but not a backward one.
Therefore, to determine whether a frame is keyframe, both bref and fref
should be confirmed to be 0.

This is a complementary commit to 7f4e47cc034df770c6792c8fc81bb0797302a001.