GUI: update check: avoid potential invalid memory access
Newer mingw gcc versions (10.3.0, 11) report that there was a
potential write beyond the end of an internal buffer in a QByteArray
instance. It is entirely unclear to me how that could happen, but it
has to do with the conversion from `constexpr auto … = "C-string"` to
`std::string` instances.

As those URLs don't really need to be available at compile time,
converting them to be returned on demand at runtime seemed fine. The
warning is gone now, but again, I don't fully understand the why &
how.