Revert "GUI: source file: use = default for copy constructor"
This reverts commit b46089387c36f3668dc2830a0808fee56634910f.

With the patch applied the default copy constructor only copies the
shared pointers to the other files & tracks & attachments — but we
need a deep copy here. Otherwise invalid memory access will happen (or
rather, the GUI's `MuxConfig::verifyStructure()` recognizes the
problem and aborts with the message "track->m_file == nullptr in level
0 file").

Fixes #3157.