Debian packaging: compile with clang on Debian 11 Bullseye
The next stable release of Debian, version 11 "Bullseye", will ship
with a gcc version that produces an "internal compiler error"
segmentation fault when compiling current versions of MKVToolNix. This
error has been fixed in gcc upstream as well as the gcc in Debian
testing, but it won't be backported to the one in Bullseye.

There are two possible workarounds:

1. Compile with clang instead (this is what this commit does)
2. Compile with gcc 9 which Bullseye also ships

As I've been using clang in parallel to gcc for years, I've opted to
compile with it instead of the older gcc 9 for Bullseye.

Ubuntu releases don't suffer from this issue:

* 20.04 comes with gcc 9.3.0
* 20.10 & 21.04 come with gcc 10.3