BCP 47: fix script/variants validation if any prefix restricts them
Example:

The `ekavsk` variant has three prefixes: `sr`, `sr-Latn` and
`sr-Cyrl`. This must be interpreted as "if the language tag contains a
script, it must be either `Latn` or `Cyrl`; no other script subtag is
allowed". Not having a script subtag is allowed as well.

In effect this means that the following three language tags are all
supposed to be valid: `sr-ekavsk`, `sr-Cyrl-ekavsk` and
`sr-Latn-ekavsk`. This part of the validation worked correctly.

However, any other script was considered to be valid as well,
e.g. `sr-Bali-ekavsk`. This was due to the first prefix `sr`, which,
on its own, is a prefix match of `sr-Bali-ekavsk`. This did not take
into account that there are other prefixes that do restrict the script
subtag.

This was fixed. Now only either no script subtag or a script subtag of
`Cyrl` or `Latn` is considered valid.