probing: prefer AVC & HEVC at start of file over audio detection
Even though AVC & HEVC are often mis-detected in the middle of other
container formats, it is pretty unambiguous if the file starts with
the typical NALU marker. So try to detect AVC & HEVC before trying
audio types if the file starts with a NALU marker as audio types are
often mis-detected as well.

Fixes #3201.