generic_video_packetizer_c: correct broken B frames handling
Back then when I authored 7f4e47cc034df770c6792c8fc81bb0797302a001,
I wasn't aware of the fact that bref of a B frame is actually the
penultimate I or P frame (with the last one being its fref; both
encoded before it).

Apparently the commit worked (as in, causing mkvmerge-remuxed file
to give identical ES stream when being mkvextract'd without --raw)
because it would work as long as bref of the B frames are not -1.
Regardless, it's probably better to make it right.

P.S. I also made a wrong statement in the commit message of
7f4e47cc034df770c6792c8fc81bb0797302a001, that m_ref_timestamp was
relevant in the fix. Now I have a clearer view on the code and I'm
pretty sure that would be impossible. (It is now relevant though.)