BCP 47: don't enforce prefixes for variants; enforce uniqueness of variants
BCP 47's verbiage is pretty lax wrt. variants & their prefixes. It
states[1]:

> Variant subtag records in the Language Subtag Registry MAY include
> one or more 'Prefix' (Section 3.1.8) fields.  Each 'Prefix'
> indicates a suitable sequence of subtags for forming (with other
> subtags, as appropriate) a language tag when using the variant.

Therefore a hard check whether a variant is used with only the listed
prefixes is inappropriate.

Furthermore there are other semi-normative sources stating the
same. For example, the W3C[2] says:

> Check the context and ordering for variant subtags. Most variant
> subtag records in the registry have one or more Prefix fields. The
> prefixes indicate with which subtags it is usually appropriate to
> use this variant.

…

> If you have a good reason, you could use a variant subtag with
> different subtags, eg. cmn-Latn-pinyin would be a perfectly legal
> way to say Mandarin Chinese written with pinyin.

And `pinyin` lists neither `cmn` nor `cmn-Latn` as a prefix.

BCP 47 goes on to state that "Most variants that share a prefix are
mutually exclusive", but there's actually no way to identify the
variants for which this holds true automatically. Therefore this
property isn't enforced either.

Lastly BCP 47 does have one hard requirement on variants in [1]:

>  5. The same variant subtag MUST NOT be used more than once within a
>     language tag.

This is now enforced.

Part of the implementation/fix of #3307.

[1]  https://www.rfc-editor.org/rfc/rfc5646.html#section-2.2.5
[2]  https://www.w3.org/International/questions/qa-choosing-language-tags#variants