GUI: BCP47: make normalization form info non-warnings depending on preferences
So far if there was a canonical and/or extlang form that was different
form what the user entered those pieces of information were shown as
warnings.

With this change the state of the preferences are taken into
account. The same pieces of information are now only a warning if the
different normalization form is the one selected in the preferences as
the one to normalize to.

For example: if you enter `yue` (which is a canonical form with the
corresponding extlang form `zh-yue`) & your preferences are set to
"canonical", then this situation is judged "OK" as the text entered
matches the form you want according to the preferences. A text with an
"info" icon is shown, stating that there's an extlang form of `zh-yue`
available.

If, however, your preferences are set to "extlang", then a text with
a "warning" icon is shown stating what the extlang form would be.

If your preferences are set to "no normalization" the messages are
always shown with an "info" icon.

Part of the implementation of #3307.