doc type version handler: don't try to make EBML head's "size" field longer than 8 bytes
When replacing an element with a new one, the new one might be exactly
one byte smaller than existing one. In that case an EBML void element
cannot be written after the new element as a void element's minimum
size is two bytes (one byte ID, one byte "size" field).

So far the strategy to deal with those was to force the new element's
"size" field to be one byte larger than it needs to be, therefore
covering that one byte after where its data would have ended had the
minimum size be used.

Problem is that the "size" field's maximum length is eight bytes. If
the existing element's "size" field is already eight bytes long,
increasing it by one won't work.

However, in the case of the "EBML head" master element this isn't a
problem as that element only ever contains a handful of
elements. Instead of making the "size" field larger, we can easily
reduce its size down to one or two bytes, freeing up up to seven bytes
— which then allows a void element to be written after the new element
as we now have up to eight bytes free after the new head element.

Part of the fix of #3325.